Practically done!  Just the title and the related
Ironically, the last part of the article that is needed is the title.  This will also affect the URL, and then the few parameters sent to Disqus and/or Google Analytics.  But it's basically done!  I will share it with friends for a review, and then publish it on Tuesday.